move setup to test_nlp.py to keep scopes better defined
